write quantized CNN model
File format is similar to export_model. I only added a few features: the number of parameters in the model, group sizes for each layer's weigths and biases, offsets for scaling factors and scaling factors to the end of file.

The group size varies for different parameters, being calculated individually for each layer's weights and biases based on their respective sizes. By default, the group size is set to 64. If the dimension of specific parameters is less than 64, the group size will be set to that dimension. If the dimension is greater than 64 but not a multiple of 64, the group size will be set to the factor of the dimension that is closest to 64.