fix batch processing and static quantization
There was a bug in batch processing in `run.c`, `runq.c` and `runq_static.c` due to the concat pooling layer. The concat pooling layer combines all inputs by passing them through a max pooling layer and an average pooling layer, then concatenating the outputs of these two layers. Before the fix, the memory allocated for the outputs of the max pooling and average pooling layers was overlapping when using batch processing. This issue has now been resolved.

Regarding static quantization, the ResNet50 model was causing the kernel to crash due to running out of memory. The issue was that I hadn't used `torch.no_grad()` in the `quantize()` function, so it was retaining all intermediate information. This has now been corrected.